Separate core modules - rocket, controls, settings
